MWPE decoder with no measurement error

+++ error_dynamics/error_model/iid_error.hpp
@@ -6,15 +6,15 @@ namespace ErrorModel {
 
 class IIDError: public ErrorModelBase{
     private:
-    float px, py, pz;
-    float pm;
+    double px, py, pz;
+    double pm;
 
     public:
     IIDError() = delete;
-    inline IIDError(float _px, float _py, float _pz, float _pm) {
+    inline IIDError(double _px, double _py, double _pz, double _pm) {
         px = _px, py = _py, pz = _pz, pm = _pm;
     };
-    inline IIDError(float p) : IIDError(p, p, p, p / 3.0f * 2.0f) {}
+    inline IIDError(double p) : IIDError(p, p, p, p / 3.0f * 2.0f) {}
 
     std::pair<std::shared_ptr<CodeScheme::RectError>, std::shared_ptr<CodeScheme::RectSyndrome>> generate_rectangular_error(const CodeScheme::RectShape shape);
